requirements.txt
pandas
openpyxl
pyarrow
pdfplumber
PyMuPDF
lxml
xmltodict
beautifulsoup4
hl7
colorlog
fastapi
uvicorn
python-multipart
sqlalchemy
python-json-logger
PyPDF2>=3.0.1
scipy
requests
xlrd
eml_parser              
loguru                  
tqdm                    
tabulate                
faiss-cpu               
pydantic                
openai
chardet
python-dotenv
